requirements.txt
alembic==1.4.2
attrs==19.3.0
Babel==2.8.0
blinker==1.4
brewer2mpl==1.4.1
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.3
Flask-Table==0.5.0
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
more-itertools==8.4.0
mpld3==0.5.1
mysqlclient==1.4.6
numpy==1.18.5
packaging==20.4
pandas==1.0.4
patsy==0.5.1
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.17
statsmodels==0.11.1
threadpoolctl==2.1.0
tzlocal==2.1
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
